pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "High-performance FastAPI template featuring Pydantic V2, Argon2 security, and Dev Container support"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard-no-fastapi-cloud-cli]>=0.135.2",
    "httptools>=0.7.1",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.12.1",
    "redis[hiredis]>=7.4.0",
    "sqlalchemy>=2.0.48",
    "user-agents>=2.2.0",
    "uvloop>=0.22.1",
]

[dependency-groups]
dev = ["locust>=2.43.3", "pre-commit>=4.5.1", "ruff>=0.15.7"]
